DownloadDirectory steps can now be run based on config file; Changes to BioPAX script to accomodate new config system

+++ downloadDirectory/src/main/java/org/reactome/release/downloadDirectory/ReactomeBookGenerator.java
@@ -14,30 +14,35 @@ import org.gk.persistence.MySQLAdaptor;
 public class ReactomeBookGenerator {
 	private static final Logger logger = LogManager.getLogger();
 	
-	static public void execute(String username, String password, String host, int port, String database, String releaseNumber, String releaseDir ) throws IOException, InterruptedException {
+	static public void execute(String username, String password, String host, int port, String database, String releaseNumber, String releaseDir, boolean runPDF, boolean runRTF) throws IOException, InterruptedException {
 		logger.info("Generating Reactome Books");
 		String pdfBookCommand = "perl " + releaseDir + "/genbook_pdf.pl -depth 100 -db " + database + " -host " + host + " -user " + username + " -pass " + password + " -port " + port + " -stdout -react_rep 2";
 		String rtfBookCommand = "perl " + releaseDir + "/genbook_rtf.pl -depth 100 -db " + database + " -host " + host + " -user " + username + " -pass " + password + " -port " + port + " -split -react_rep 2";
 		
-		logger.info("Generating PDF Reactome Book ...");
-		Process generatePDFBook = Runtime.getRuntime().exec(pdfBookCommand);
-		generatePDFBook.waitFor();
-		
-		Process zipPDFBook = Runtime.getRuntime().exec("zip TheReactomeBook.pdf.zip TheReactomeBook.pdf");
-		zipPDFBook.waitFor();
-		
-		Files.move(Paths.get("TheReactomeBook.pdf.zip"), Paths.get(releaseNumber + "/TheReactomeBook.pdf.zip"), StandardCopyOption.REPLACE_EXISTING); 
-		Files.delete(Paths.get("TheReactomeBook.pdf"));
+		if (runPDF) {
+			logger.info("Generating PDF Reactome Book ...");
+			Process generatePDFBook = Runtime.getRuntime().exec(pdfBookCommand);
+			generatePDFBook.waitFor();
+			
+			Process zipPDFBook = Runtime.getRuntime().exec("zip TheReactomeBook.pdf.zip TheReactomeBook.pdf");
+			zipPDFBook.waitFor();
+			
+			Files.move(Paths.get("TheReactomeBook.pdf.zip"), Paths.get(releaseNumber + "/TheReactomeBook.pdf.zip"), StandardCopyOption.REPLACE_EXISTING); 
+			Files.delete(Paths.get("TheReactomeBook.pdf"));
+			logger.info("Finished generating PDF Reactome Book");
+		}
 
-		logger.info("Generating RTF Reactome Book ...");
-		Process generateRTFBook = Runtime.getRuntime().exec(rtfBookCommand);
-		generateRTFBook.waitFor();
-		
-		Process zipRTFBook = Runtime.getRuntime().exec("zip TheReactomeBook.rtf.zip TheReactomeBook");
-		zipRTFBook.waitFor();
-		Files.move(Paths.get("TheReactomeBook.rtf.zip"), Paths.get(releaseNumber + "/TheReactomeBook.rtf.zip"), StandardCopyOption.REPLACE_EXISTING); 
-		FileUtils.deleteDirectory(new File("TheReactomeBook"));
-		
-		logger.info("Finished generating Reactome Books");
+		if (runRTF) {
+			logger.info("Generating RTF Reactome Book ...");
+			Process generateRTFBook = Runtime.getRuntime().exec(rtfBookCommand);
+			generateRTFBook.waitFor();
+			
+			Process zipRTFBook = Runtime.getRuntime().exec("zip TheReactomeBook.rtf.zip TheReactomeBook");
+			zipRTFBook.waitFor();
+			
+			Files.move(Paths.get("TheReactomeBook.rtf.zip"), Paths.get(releaseNumber + "/TheReactomeBook.rtf.zip"), StandardCopyOption.REPLACE_EXISTING); 
+			FileUtils.deleteDirectory(new File("TheReactomeBook"));
+			logger.info("Finished generating RTF Reactome Book");
+		}
 	}
 }
